--- DiscordChatExporter.Core.Markdown/Node.cs
@@ -1,12 +1,0 @@
-﻿namespace DiscordChatExporter.Core.Markdown
-{
-    public abstract class Node
-    {
-        public string Lexeme { get; }
-
-        protected Node(string lexeme)
-        {
-            Lexeme = lexeme;
-        }
-    }
-}
