+++ DiscordChatExporter.Core.Markdown/Nodes/MentionType.cs
@@ -1,4 +1,4 @@
-﻿namespace DiscordChatExporter.Core.Markdown
+﻿namespace DiscordChatExporter.Core.Markdown.Nodes
 {
     public enum MentionType
     {
