+++ DiscordChatExporter.Core.Markdown/Nodes/Node.cs
@@ -1,0 +1,12 @@
+﻿namespace DiscordChatExporter.Core.Markdown.Nodes
+{
+    public abstract class Node
+    {
+        public string Source { get; }
+
+        protected Node(string source)
+        {
+            Source = source;
+        }
+    }
+}
