+++ DiscordChatExporter.Core.Markdown/Nodes/TextFormatting.cs
@@ -1,4 +1,4 @@
-﻿namespace DiscordChatExporter.Core.Markdown
+﻿namespace DiscordChatExporter.Core.Markdown.Nodes
 {
     public enum TextFormatting
     {
